+++ src/python/tests/test_kmer_map.py
@@ -1,0 +1,8 @@
+import oxli
+import pytest
+from test_basic import create_sample_kmer_table
+
+
+def test_kmermap():
+    '''Test option to add kmermap'''
+    pass
