+++ src/python/tests/test_output.py
@@ -1,0 +1,24 @@
+import oxli
+import pytest
+from test_basic import create_sample_kmer_table
+
+
+def test_serialise():
+    '''Serialise object to JSON '''
+    pass
+
+def test_deserialise():
+    '''Load object from file.'''
+    pass
+
+def test_dump():
+    '''Write tab delimited kmer:count pairs'''
+    pass
+
+def test_dump_hash():
+    '''Write tab delimited hash_count pairs '''
+    pass
+
+def test_histo():
+    '''Write frequency counts.'''
+    pass
