+++ src/python/tests/test_remove.py
@@ -1,0 +1,19 @@
+import oxli
+import pytest
+from test_basic import create_sample_kmer_table
+
+def test_drop():
+    '''Remove kmer by name.'''
+    pass
+
+def test_drop_hash():
+    '''Remove record by hash.'''
+    pass
+
+def test_mincut():
+    '''Remove all records with counts < threshold. '''
+    pass
+
+def test_maxcut():
+    '''Remove all records with counts > threshold. '''
+    pass
